Create llprint.h
General code cleanup
